♻️ move evaluation tests
